% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEtype.R
\docType{data}
\name{counts}
\alias{counts}
\title{counts: A test data for DEsinge}
\format{A non-negative integer matrix of scRNA-seq read counts, rows are genes and columns are samples.}
\source{
Petropoulos S, et al. Cell, 2016, 165(4): 1012-1026.
}
\usage{
data(TestData)
}
\description{
A matrix of read counts of single-cell RNA-seq data for testing of DEsingle.
}
\examples{
# Load library and the test data for DEsingle
library(DEsingle)
data(TestData)

# Specifying the two groups to be compared
# The sample number in group 1 and group 2 is 50 and 100 respectively
group <- factor(c(rep(1,50), rep(2,100)))

# Detecting the differentially expressed genes
results <- DEsingle(counts = counts, group = group)

# Dividing the differentially expressed genes into 3 categories
results.classified <- DEtype(results = results, threshold = 0.05)

}
\seealso{
\code{\link{DEsingle}}, for the detection of differentially expressed genes from scRNA-seq data.

\code{\link{DEtype}}, for the classification of differentially expressed genes found by \code{\link{DEsingle}}.

\code{\link{counts}} and \code{\link{group}}, a test data for DEsingle.
}
\keyword{data}
