% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestData.R
\docType{data}
\name{TestData}
\alias{TestData}
\alias{counts}
\alias{group}
\title{TestData: A test dataset for DEsinge}
\format{\itemize{
  \item counts. A non-negative integer matrix of scRNA-seq read counts, rows are genes and columns are cells.
  \item group. A vector of factor specifying the two groups to be compared, corresponding to the column of the \code{counts}. Also could be generated by: group <- factor(c(rep(1,50), rep(2,100)))
}}
\source{
Petropoulos S, et al. Cell, 2016, 165(4): 1012-1026.
}
\usage{
data(TestData)
}
\description{
A toy dataset containing a single-cell RNA-seq read counts matrix and its grouping information.
}
\details{
\itemize{
  \item counts. A matrix of raw read counts of single-cell RNA-seq data which has 200 genes (rows) and 150 cells (columns).
  \item group. A vector of factor specifying the two groups to be compared in \code{counts}.
}
}
\examples{
# Load test data for DEsingle
data(TestData)

# Specifying the two groups to be compared
# The sample number in group 1 and group 2 is 50 and 100 respectively
group <- factor(c(rep(1,50), rep(2,100)))

# Detecting the differentially expressed genes
results <- DEsingle(counts = counts, group = group)

# Dividing the differentially expressed genes into 3 categories
results.classified <- DEtype(results = results, threshold = 0.05)

}
\seealso{
\code{\link{DEsingle}}, for the detection of differentially expressed genes from scRNA-seq data.

\code{\link{DEtype}}, for the classification of differentially expressed genes found by \code{\link{DEsingle}}.
}
\keyword{data}
